requirements.txt
attrs==20.3.0
autopep8==1.5.6
base58check==1.0.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.23.0
pycodestyle==2.7.0
pymongo==3.11.3
pysha3==1.0.2
python-dotenv==0.17.0
requests==2.25.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.4.0
